refactor: rename config directory to logging
