+++ lib/logging/logger.js
