+++ lib/middleware/loggingHandler.js
@@ -1,4 +1,4 @@
-const winston = require('../config/logger')
+const winston = require('../logging/logger')
 const expressWinston = require('express-winston')
 const jwt = require('jsonwebtoken')
