+++ lib/utils/circuitBreaker.js
@@ -1,5 +1,5 @@
 const request = require('request-promise')
-const LOGGER = require('../config/logger')
+const LOGGER = require('../logging/logger')
 
 /**
  * Implementation of "circuit breaker" pattern: retry + sleep + temporary deactivation
